+++ vck-openid/src/commonTest/kotlin/at/asitplus/wallet/lib/oidvci/OidvciOfferCodeTest.kt
@@ -106,7 +106,7 @@ val OidvciOfferCodeTest by testSuite {
             val credentialIdToRequest = it.mapper.toCredentialIdentifier(AtomicAttribute2023, PLAIN_JWT)
             val credentialOffer = it.authorizationService.credentialOfferWithAuthorizationCode(
                 credentialIssuer = it.issuer.publicContext,
-                configurationIds = setOf(credentialIdToRequest)
+                credentials = setOf(AtomicAttribute2023 to PLAIN_JWT)
             )
             val credentialFormat = it.issuer.metadata.supportedCredentialConfigurations!![credentialIdToRequest]
                 .shouldNotBeNull()
@@ -134,7 +134,7 @@ val OidvciOfferCodeTest by testSuite {
         test("process with code after credential offer, and scope for all credentials") {
             val credentialOffer = it.authorizationService.credentialOfferWithAuthorizationCode(
                 credentialIssuer = it.issuer.publicContext,
-                configurationIds = listOf(),
+                credentials = emptySet(),
             )
             val credentialIdToRequest = credentialOffer.configurationIds.first()
             val credentialFormat = it.issuer.metadata.supportedCredentialConfigurations!![credentialIdToRequest]
@@ -163,7 +163,7 @@ val OidvciOfferCodeTest by testSuite {
         test("process with code after credential offer, wrong issuer_state") {
             val credentialOffer = it.authorizationService.credentialOfferWithAuthorizationCode(
                 credentialIssuer = it.issuer.publicContext,
-                configurationIds = listOf(),
+                credentials = emptySet(),
             )
             val credentialIdToRequest = credentialOffer.configurationIds.first()
             val credentialFormat =
@@ -184,7 +184,7 @@ val OidvciOfferCodeTest by testSuite {
         test("process with code after credential offer over par and request_uri") {
             val credentialOffer = it.authorizationService.credentialOfferWithAuthorizationCode(
                 credentialIssuer = it.issuer.publicContext,
-                configurationIds = listOf(),
+                credentials = emptySet(),
             )
             val credentialIdToRequest = credentialOffer.configurationIds.first()
             val credentialFormat = it.issuer.metadata.supportedCredentialConfigurations
@@ -210,7 +210,7 @@ val OidvciOfferCodeTest by testSuite {
         test("process with code after credential offer, and authorization details for one credential") {
             val credentialOffer = it.authorizationService.credentialOfferWithAuthorizationCode(
                 credentialIssuer = it.issuer.publicContext,
-                configurationIds = listOf(),
+                credentials = emptySet(),
             )
             val credentialIdToRequest = credentialOffer.configurationIds.first()
             val authorizationDetails = it.client.buildAuthorizationDetails(
@@ -246,7 +246,7 @@ val OidvciOfferCodeTest by testSuite {
             val credentialIdToRequest = it.mapper.toCredentialIdentifier(AtomicAttribute2023, PLAIN_JWT)
             val credentialOffer = it.authorizationService.credentialOfferWithAuthorizationCode(
                 credentialIssuer = it.issuer.publicContext,
-                configurationIds = setOf(credentialIdToRequest)
+                credentials = setOf(AtomicAttribute2023 to PLAIN_JWT)
             )
             val credentialFormat = it.issuer.metadata.supportedCredentialConfigurations!![credentialIdToRequest]
                 .shouldNotBeNull()
@@ -265,7 +265,7 @@ val OidvciOfferCodeTest by testSuite {
             val credentialIdToRequest = it.mapper.toCredentialIdentifier(AtomicAttribute2023, PLAIN_JWT)
             val credentialOffer = it.authorizationService.credentialOfferWithAuthorizationCode(
                 credentialIssuer = it.issuer.publicContext,
-                configurationIds = setOf(credentialIdToRequest)
+                credentials = setOf(AtomicAttribute2023 to PLAIN_JWT)
             )
             val otherCredentialIdToRequest = it.mapper.toCredentialIdentifier(AtomicAttribute2023, ISO_MDOC)
             // important step: mess with the scope value
@@ -280,7 +280,7 @@ val OidvciOfferCodeTest by testSuite {
             val credentialIdToRequest = it.mapper.toCredentialIdentifier(AtomicAttribute2023, PLAIN_JWT)
             val credentialOffer = it.authorizationService.credentialOfferWithAuthorizationCode(
                 credentialIssuer = it.issuer.publicContext,
-                configurationIds = setOf(credentialIdToRequest)
+                credentials = setOf(AtomicAttribute2023 to PLAIN_JWT)
             )
             val otherCredentialIdToRequest = it.mapper.toCredentialIdentifier(AtomicAttribute2023, ISO_MDOC)
             // important step: mess with the authorization details